fix consecutive string literals bug
adding space separators in the external scanner introduced a bug with
consecutive string literals.
i.e `"a string" " and " "another one"` is a valid sc expression.

This is because there was no rules allowing a space separator between
literals. It was not a problem before because the spaces were handled by
the `extras` rules.
Allowing repetitions of string literals seems to fix it, but treat all
consecutive strings as one.